+++ game_patch/graphics/gr.h
@@ -13,6 +13,7 @@ void gr_bitmap_scaled_float(int bitmap_handle, float x, float y, float w, float 
 float gr_scale_fov_hor_plus(float horizontal_fov);
 bool gr_3d_bitmap_oriented_wh(const rf::Vector3* pnt, const rf::Matrix3* M, float half_w, float half_h, rf::gr::Mode mode);
 int gr_fit_string(std::string& text, int max_width, int font_id = -1, std::string_view suffix = "-");
+void explosion_flash_lights_level_init();
 
 inline constexpr rf::gr::Mode overdraw_colorized_3d_bitmap{
         rf::gr::TEXTURE_SOURCE_WRAP,