+++ game_patch/misc/player.cpp
@@ -512,6 +512,26 @@ ConsoleCommand2 damage_screen_flash_cmd{
     "Toggle damage screen flash effect",
 };
 
+ConsoleCommand2 weapon_explosion_flash_lights_cmd{
+    "cl_explosionflashweapons",
+    []() {
+        g_alpine_game_config.explosion_weapon_flash_lights = !g_alpine_game_config.explosion_weapon_flash_lights;
+        rf::console::print("Explosion flash lights from weapons are {}",
+            g_alpine_game_config.explosion_weapon_flash_lights ? "enabled" : "disabled");
+    },
+    "Toggle explosion flash dynamic lights from weapons",
+};
+
+ConsoleCommand2 env_explosion_flash_lights_cmd{
+    "cl_explosionflashenv",
+    []() {
+        g_alpine_game_config.explosion_env_flash_lights = !g_alpine_game_config.explosion_env_flash_lights;
+        rf::console::print("Explosion flash lights from environmental sources are {}",
+            g_alpine_game_config.explosion_env_flash_lights ? "enabled" : "disabled");
+    },
+    "Toggle explosion flash dynamic lights from environmental sources",
+};
+
 void play_local_sound_3d(uint16_t sound_id, rf::Vector3 pos, int group, float volume) {
     rf::snd_play_3d(sound_id, pos, volume, rf::Vector3{}, group);
 }
@@ -856,6 +876,8 @@ void player_do_patch()
 
     // Commands
     damage_screen_flash_cmd.register_cmd();
+    weapon_explosion_flash_lights_cmd.register_cmd();
+    env_explosion_flash_lights_cmd.register_cmd();
     death_bars_cmd.register_cmd();
     swap_assault_rifle_controls_cmd.register_cmd();
     swap_grenade_controls_cmd.register_cmd();