+++ docs/CHANGELOG.md
@@ -9,6 +9,8 @@ Version 1.3.0 (TBD): Not yet released
 [@GooberRF](https://github.com/GooberRF)
 - Add values to scope param for `Scope_Gate` event to test for team regardless of whether local player is spawned
 - Add `When_Round_Ends` event
+- Add dynamic light flashes for explosions from weapons and clutter/entity deaths
+- Add `cl_explosionflashweapons` and `cl_explosionflashenv` commands
 
 ### Bug fixes
 [@GooberRF](https://github.com/GooberRF)